Fixing ProjectBase.files paths and ProjectDesignerDocumentManager (show project properties)
git-svn-id: https://MPFProj10.svn.codeplex.com/svn@50908 81fbe566-5dc4-48c1-bdea-7421811ca204

+++ Dev10/Src/CSharp/ProjectDesignerDocumentManager.cs
@@ -55,9 +55,9 @@ namespace Microsoft.VisualStudio.Project
 
 				IVsUIHierarchy parentHierarchy = this.Node.ProjectMgr.GetProperty((int)__VSHPROPID.VSHPROPID_ParentHierarchy) as IVsUIHierarchy;
 
-				uint parentHierarchyItemId = (uint)this.Node.ProjectMgr.GetProperty((int)__VSHPROPID.VSHPROPID_ParentHierarchyItemid);
+                int parentHierarchyItemId = (int)this.Node.ProjectMgr.GetProperty((int)__VSHPROPID.VSHPROPID_ParentHierarchyItemid);
 
-				ErrorHandler.ThrowOnFailure(uiShellOpenDocument.OpenSpecificEditor(editorFlags, fullPath, ref editorType, physicalView, ref logicalView, caption, parentHierarchy, parentHierarchyItemId, docDataExisting, serviceProvider, out frame));
+                ErrorHandler.ThrowOnFailure(uiShellOpenDocument.OpenSpecificEditor(editorFlags, fullPath, ref editorType, physicalView, ref logicalView, caption, parentHierarchy, (uint)parentHierarchyItemId, docDataExisting, serviceProvider, out frame));
 
 				if(frame != null)
 				{
